feat: add createFilePath function to utilities
